enhancement: add mimetype to file extension rego function
add rego function to detect the resource extension by mimetype, at the same time this pr introduces a custom ocis namespace for the rego functions.
